[Decomp][ObjectMgr] Cohesion split: extract trainer/vendor loaders to ObjectMgrTrainerVendor.cpp
